removed hook and made normal helper function
